+++ lizard_damage/models.py
@@ -8,16 +8,23 @@ from __future__ import (
 )
 from django.contrib.gis.db import models
 from django.core.urlresolvers import reverse
+from django.core.files import File
 from lizard_map import coordinates
 from lizard_task.models import SecuredPeriodicTask
 from pyproj import transform
 from pyproj import Proj
+import matplotlib as mpl
+import numpy as np
 
 import datetime
 import os
 import random
 import string
 import json
+import tempfile
+import Image
+import subprocess
+from osgeo import gdal
 
 # from django.utils.translation import ugettext_lazy as _
 
@@ -29,6 +36,37 @@ WGS84 = str('+proj=latlong +datum=WGS84')
 
 rd_proj = Proj(RD)
 wgs84_proj = Proj(WGS84)
+
+
+def extent_from_geotiff(filename):
+    ds = gdal.Open(filename)
+    width = ds.RasterXSize
+    height = ds.RasterYSize
+    gt = ds.GetGeoTransform()
+    minx = gt[0]
+    miny = gt[3] + width*gt[4] + height*gt[5]
+    maxx = gt[0] + width*gt[1] + height*gt[2]
+    maxy = gt[3]
+    return (minx, miny, maxx, maxy)
+
+
+def write_pgw(name, extent):
+    """write pgw file:
+
+    0.5
+    0.000
+    0.000
+    0.5
+    <x ul corner>
+    <y ul corner>
+
+    extent is a 4-tuple
+    """
+    f = open(name, 'w')
+    f.write('0.5\n0.000\n0.000\n-0.5\n')
+    f.write('%f\n%f' % (min(extent[0], extent[2]), max(extent[1], extent[3])))
+    f.close()
+    return
 
 
 def friendly_filesize(size):
@@ -49,7 +87,7 @@ def friendly_filesize(size):
 class AhnIndex(models.Model):
     """
     Sql for this table can be generated using:
-    
+
     shp2pgsql -s 28992 ahn2_05_int_index public.data_index | sed\
     s/geom/the_geom/g > index.sql
 
@@ -176,10 +214,11 @@ class DamageScenario(models.Model):
         (0, '1 Kaart met de max waterstand van 1 gebeurtenis'),
         (1, '1 Kaart met de waterstand voor een zekere herhalingstijd'),
         (2, 'Kaarten met per tijdstip de waterstand van 1 gebeurtenis'),
-        (3, 'Kaarten met de max. waterstand van afzonderlijke gebeurtenissen.'),
-        (4, 'Kaarten met voor verschillende herhalingstijden de waterstanden'),
+        (3, 'Kaarten met de max. waterstand van afzonderlijke gebeurtenissen'),
+        (4, 'Kaarten met voor verschillende herhalingstijden de waterstanden (voor risicokaart)'),
         (5, 'Tijdserie aan kaarten met per tijdstip de '
             'waterstand van meerdere gebeurtenissen'),
+        (6, 'Batenkaart maken met resultaten uit twee risicokaarten'),
     )
     SCENARIO_TYPES_DICT = dict(SCENARIO_TYPES)
 
@@ -273,6 +312,14 @@ class DamageEvent(models.Model):
         upload_to='scenario/result',
         null=True, blank=True,
         help_text='Will be filled once the calculation has been done')
+    landuse_slugs = models.TextField(
+        null=True, blank=True,
+        help_text='comma separated landuse slugs for GeoImage')
+    height_slugs = models.TextField(
+        null=True, blank=True,
+        help_text='comma separated height slugs for GeoImage')
+    min_height = models.FloatField(null=True, blank=True)
+    max_height = models.FloatField(null=True, blank=True)
 
     def __unicode__(self):
         if self.name: return self.name
@@ -343,3 +390,173 @@ class DamageEventWaterlevel(models.Model):
 
     def __unicode__(self):
         return os.path.basename(self.waterlevel.path)
+
+
+class BenefitScenario(models.Model):
+    """Baten berekening.
+
+    Results will be in result_zip and
+    in BenefitScenarioResult.
+    """
+    name = models.CharField(max_length=64)
+    slug = models.SlugField(null=True, blank=True, help_text='auto generated on save; used for url')
+    email = models.EmailField(max_length=128)
+
+    datetime_created = models.DateTimeField(auto_now=True)
+    expiration_date = models.DateTimeField()
+
+    zip_risk_a = models.FileField(upload_to='benefit/risk')
+    zip_risk_b = models.FileField(upload_to='benefit/risk')
+
+    zip_result = models.FileField(
+        upload_to='benefit/result',
+        null=True, blank=True,
+        help_text='Will be filled when results are available')
+
+    def __unicode__(self):
+        return '%s' % (self.name)
+
+    def save(self, *args, **kwargs):
+        if not self.slug:
+            self.slug = ''.join(random.sample(string.letters, 20))
+        if not self.expiration_date:
+            self.expiration_date = datetime.datetime.now() + datetime.timedelta(days=7)
+        return super(BenefitScenario, self).save(*args, **kwargs)
+
+    def get_absolute_url(self):
+        return reverse('lizard_damage_benefit_result', kwargs=dict(slug=self.slug))
+
+    def result_display(self):
+        try:
+            return 'zipfile (%s)' % friendly_filesize(self.zip_result.size)
+        except:
+            return 'geen zipfile'
+
+
+class BenefitScenarioResult(models.Model):
+    """ Result of 1 tile of a Benefit Scenario
+
+    Used to create kml
+    """
+    benefit_scenario = models.ForeignKey(BenefitScenario)
+    image = models.FileField(upload_to='scenario/image')
+
+    north = models.FloatField()
+    south = models.FloatField()
+    east = models.FloatField()
+    west = models.FloatField()
+
+    def __unicode__(self):
+        return '%s - %s' % (self.benefit_scenario, self.image)
+
+
+class GeoImage(models.Model):
+    """
+    Generic geo referenced image
+
+    i.e. For use in kml files
+    """
+    slug = models.SlugField()
+    image = models.FileField(upload_to='geoimage')
+
+    north = models.FloatField()
+    south = models.FloatField()
+    east = models.FloatField()
+    west = models.FloatField()
+
+    def __unicode__(self):
+        return self.slug
+
+    @classmethod
+    def from_data_with_legend(cls, slug, data, extent, legend):
+        """
+        Create GeoImage from slug and data.
+
+        Data is numpy array.
+        """
+        tmp_base = tempfile.mktemp()
+        #print('tmp_base: %s' % tmp_base)
+        #print('step 1')
+        # Step 1: save png + pgw in RD
+
+        colormap = mpl.colors.ListedColormap(legend, 'indexed')
+        rgba = colormap(data, bytes=True)
+        #rgba[:,:,3] = np.where(rgba[:,:,0], 153 , 0)
+        Image.fromarray(rgba).save(tmp_base + '.png', 'PNG')
+        write_pgw(tmp_base + '.pgw', extent)
+
+        return cls.from_rd_png(tmp_base, slug, extent)
+
+
+    @classmethod
+    def from_data_with_min_max(cls, slug, data, extent, min_value, max_value):
+        """
+        Create GeoImage from slug and data.
+        """
+        tmp_base = tempfile.mktemp()
+        #print('tmp_base: %s' % tmp_base)
+        #print('step 1')
+        # Step 1: save png + pgw in RD
+
+        cdict = {
+            'red': ((0.0, 51./256, 51./256),
+                    (0.5, 237./256, 237./256),
+                    (1.0, 83./256, 83./256)),
+            'green': ((0.0, 114./256, 114./256),
+                      (0.5, 245./256, 245./256),
+                      (1.0, 83./256, 83./256)),
+            'blue': ((0.0, 54./256, 54./256),
+                     (0.5, 170./256, 170./256),
+                     (1.0, 83./256, 83./256)),
+            }
+        colormap = mpl.colors.LinearSegmentedColormap('something', cdict, N=1024)
+        normalize = mpl.colors.Normalize(vmin=min_value, vmax=max_value)
+        rgba = colormap(normalize(data), bytes=True)
+        #rgba[:,:,3] = np.where(rgba[:,:,0], 153 , 0)
+        Image.fromarray(rgba).save(tmp_base + '.png', 'PNG')
+
+        write_pgw(tmp_base + '.pgw', extent)
+
+        return cls.from_rd_png(tmp_base, slug, extent)
+
+
+    @classmethod
+    def from_rd_png(cls, tmp_base, slug, extent):
+        """
+        Input: <tmp_base>.png
+        Output: geo_image object
+        """
+
+        # Step 2: warp using gdalwarp to lon/lat in .tif
+        #logger.info('Warping png to tif... %s' % img['filename_png'])
+        command = 'gdalwarp %s %s -t_srs "+proj=latlong +datum=WGS83" -s_srs "%s"' % (
+            tmp_base + '.png', tmp_base + '.tif', RD.strip())
+        #logger.info(command)
+        # Warp png file, output is tif.
+        subprocess.call([
+                'gdalwarp', tmp_base + '.png', tmp_base + '.tif',
+                '-t_srs', "+proj=latlong +datum=WGS84", '-s_srs', RD.strip()])
+
+        result_extent = extent_from_geotiff(tmp_base + '.tif')
+
+        # Step 3: convert .tif back to .png
+        im = Image.open(tmp_base + '.tif')
+        im.save(tmp_base + '_2.png', 'PNG')
+
+        # Step 4: put .png in GeoObject, with new extent
+        geo_image = cls(slug=slug)
+        geo_image.north = result_extent[3]
+        geo_image.south = result_extent[1]
+        geo_image.east = result_extent[2]
+        geo_image.west = result_extent[0]
+        with open(tmp_base + '_2.png', 'rb') as img_file:
+            geo_image.image.save(slug + '.png', File(img_file), save=True)
+        geo_image.save()
+
+        # Step 5: clean tempfiles.
+        os.remove(tmp_base + '.png')
+        os.remove(tmp_base + '.pgw')
+        os.remove(tmp_base + '.tif')
+        os.remove(tmp_base + '_2.png')
+
+        return geo_image