+++ lizard_damage/tasks.py
@@ -5,9 +5,12 @@ from django.template import Context
 #from django.template import Template
 from django.template.loader import get_template
 
+from lizard_damage.models import BenefitScenario
+from lizard_damage.models import BenefitScenarioResult
 from lizard_damage.models import DamageScenario
 from lizard_damage.models import DamageEventResult
 from lizard_damage.models import RD
+from lizard_damage.models import extent_from_geotiff
 from lizard_damage import calc
 from lizard_task.task import task_logging
 from lizard_task.models import SecuredPeriodicTask
@@ -26,18 +29,6 @@ from osgeo import gdal
 from PIL import Image
 
 
-def extent_from_geotiff(filename):
-    ds = gdal.Open(filename)
-    width = ds.RasterXSize
-    height = ds.RasterYSize
-    gt = ds.GetGeoTransform()
-    minx = gt[0]
-    miny = gt[3] + width*gt[4] + height*gt[5] 
-    maxx = gt[0] + width*gt[1] + height*gt[2]
-    maxy = gt[3] 
-    return (minx, miny, maxx, maxy)
-
-
 def convert_tif_to_png(filename_tif, filename_png):
     im = Image.open(filename_tif)
     im.save(filename_png, 'PNG')
@@ -47,7 +38,7 @@ def damage_scenario_to_task(damage_scenario, username="admin"):
     """
     Send provided damage scenario as task
     """
-    task_name = 'Calculate damage scenario %d' % damage_scenario.id
+    task_name = 'Scenario (%05d) calculate damage' % damage_scenario.id
     task_kwargs = '{"username": "%s", "taskname": "%s", "damage_scenario_id": "%d"}' % (
         username, task_name, damage_scenario.id)
     calc_damage_task, created = SecuredPeriodicTask.objects.get_or_create(
@@ -60,11 +51,28 @@ def damage_scenario_to_task(damage_scenario, username="admin"):
     calc_damage_task.send_task(username=username)
 
 
+def benefit_scenario_to_task(benefit_scenario, username="admin"):
+    """
+    Send provided benefit scenario as task
+    """
+    task_name = 'Scenario (%05d) calculate benefit' % benefit_scenario.id
+    task_kwargs = '{"username": "%s", "taskname": "%s", "benefit_scenario_id": "%d"}' % (
+        username, task_name, benefit_scenario.id)
+    calc_damage_task, created = SecuredPeriodicTask.objects.get_or_create(
+        name=task_name, defaults={
+            'kwargs': task_kwargs,
+            'task': 'lizard_damage.tasks.calculate_benefit'
+            })
+    calc_damage_task.task = 'lizard_damage.tasks.calculate_benefit'
+    calc_damage_task.save()
+    calc_damage_task.send_task(username=username)
+
+
 def send_email_to_task(damage_scenario_id, mail_template, subject, username='admin', email=""):
     """
     Create a task for sending email
     """
-    task_name = 'Send %s mail for scenario %d' % (mail_template, damage_scenario_id)
+    task_name = 'Scenario (%05d) send mail %s' % (damage_scenario_id, mail_template)
     task_kwargs = '{"username": "admin", "taskname": "%s", "damage_scenario_id": "%d", "mail_template": "%s", "subject": "%s", "email": "%s"}' % (task_name, damage_scenario_id, mail_template, subject, email)
     email_task, created = SecuredPeriodicTask.objects.get_or_create(
         name=task_name, defaults={
@@ -184,6 +192,8 @@ def calculate_damage(damage_scenario_id, username=None, taskname=None, loglevel=
 
             # result[2] is the table in a data structure
             damage_event.table = json.dumps(result[2])
+            damage_event.landuse_slugs = ','.join(result[3])  # Store references to GeoImage objects
+            damage_event.height_slugs = ','.join(result[4])  # Store references to GeoImage objects
             damage_event.save()
 
             # result[1] is a list of png files to be uploaded to the django db.
@@ -202,13 +212,13 @@ def calculate_damage(damage_scenario_id, username=None, taskname=None, loglevel=
                 logger.info(command)
                 # Warp png file, output is tif.
                 subprocess.call([
-                    'gdalwarp', img['filename_png'], img['filename_tif'], 
+                    'gdalwarp', img['filename_png'], img['filename_tif'],
                     '-t_srs', "+proj=latlong +datum=WGS84", '-s_srs', RD.strip()])
 
                 img['extent'] = extent_from_geotiff(img['filename_tif'])
                 # Convert it back to png
                 #subprocess.call([
-                #    'convert', img['filename_tif'], img['filename_png']]) 
+                #    'convert', img['filename_tif'], img['filename_png']])
                 convert_tif_to_png(img['filename_tif'], img['filename_png'])
 
             #logger.info('Creating damage event results...')
@@ -245,6 +255,21 @@ def calculate_damage(damage_scenario_id, username=None, taskname=None, loglevel=
         subject = 'STOWA Schade Calculator: scenario %s heeft fouten' % damage_scenario.name
         send_email_to_task(damage_scenario.id, 'email_error', subject, username=username)
         send_email_to_task(damage_scenario.id, 'email_error', subject, username=username,
-                           email='jack.ha@nelen-schuurmans.nl')
+                           email='olivier.hoes@nelen-schuurmans.nl')
         logger.info("finished with errors")
         return 'failure'
+
+
+@task
+@task_logging
+def calculate_benefit(benefit_scenario_id, username=None, taskname=None, loglevel=20):
+    logger = logging.getLogger(taskname)
+    logger.info("calculate benefit")
+    benefit_scenario = BenefitScenario.objects.get(pk=benefit_scenario_id)
+    logger.info("scenario: %d, %s" % (benefit_scenario.id, str(benefit_scenario)))
+
+    # TODO: make zipfile to add to benefit_scenario.zip_result
+    # and add BenefitScenarioResult objects.
+    logger.info("Not yet implemented")
+
+    return 'failure'