+++ lizard_damage/views.py
@@ -4,27 +4,35 @@ from __future__ import unicode_literals
 from django.utils.translation import ugettext as _
 from django.core.files import File
 from django.core.urlresolvers import reverse
+from django.http import HttpResponse
 # from lizard_map.views import MapView
 from django.views.generic import TemplateView
+from django.views.generic import View
 from django.shortcuts import get_object_or_404
 from django.shortcuts import render_to_response
 from django.template import RequestContext
 from django.core.files.storage import FileSystemStorage
 from django.contrib.sites.models import Site
+from django.conf import settings
 
 from lizard_damage import tasks
+from lizard_damage.models import BenefitScenario
 from lizard_damage.models import DamageScenario
 from lizard_damage.models import DamageEvent
 from lizard_damage.models import DamageEventWaterlevel
+from lizard_damage.models import GeoImage
 from lizard_ui.views import ViewContextMixin
 from lizard_damage import tools
 from lizard_damage import forms
 
+from zipfile import ZipFile
 import datetime
 import tempfile
 import os
 import re
-from zipfile import ZipFile
+from PIL import Image
+from PIL import ImageDraw
+from PIL import ImageFont
 
 temp_storage_location = tempfile.mkdtemp()
 temp_storage = FileSystemStorage(location=temp_storage_location)
@@ -57,6 +65,7 @@ DamageScenario types:
 4, 'Kaarten met voor verschillende herhalingstijden de waterstanden'),
 5, 'Tijdserie aan kaarten met per tijdstip de '
    'waterstand van meerdere gebeurtenissen'),
+6, 'Batenkaart',
 """
 def damage_scenario_from_type_0(all_form_data):
     damage_scenario = DamageScenario(
@@ -67,7 +76,7 @@ def damage_scenario_from_type_0(all_form_data):
         damage_scenario.damagetable=all_form_data['damagetable']
     damage_scenario.save()
     repairtime_roads = float(all_form_data['repairtime_roads']) * 3600 * 24
-    repairtime_buildings = float(all_form_data['repairtime_roads']) * 3600 * 24
+    repairtime_buildings = float(all_form_data['repairtime_buildings']) * 3600 * 24
     damage_event = damage_scenario.damageevent_set.create(
         floodtime=all_form_data['floodtime'] * 3600,
         repairtime_roads=repairtime_roads,
@@ -89,7 +98,7 @@ def damage_scenario_from_type_1(all_form_data):
         damage_scenario.damagetable=all_form_data['damagetable']
     damage_scenario.save()
     repairtime_roads = float(all_form_data['repairtime_roads']) * 3600 * 24
-    repairtime_buildings = float(all_form_data['repairtime_roads']) * 3600 * 24
+    repairtime_buildings = float(all_form_data['repairtime_buildings']) * 3600 * 24
     damage_event = damage_scenario.damageevent_set.create(
         repetition_time=all_form_data['repetition_time'],  # Difference is here
         floodtime=all_form_data['floodtime'] * 3600,
@@ -134,7 +143,7 @@ def unpack_zipfile_into_scenario(zipfile, scenario_name='', scenario_email=''):
                 scenario_data['calc_type'] = {'min': 1, 'max': 2, 'avg': 3}.get(line[1].lower(), 'max')
             elif line[0] == 'scenario_damage_table':
                 if line[1]:
-                    zip_temp = tempfile.mktemp()
+                    zip_temp = tempfile.mkdtemp()
                     myzip.extract(line[1], zip_temp)  # extract to temp dir
                     damage_table = os.path.join(zip_temp, line[1])
             elif line[0] == 'event_name':
@@ -175,7 +184,7 @@ def unpack_zipfile_into_scenario(zipfile, scenario_name='', scenario_email=''):
                     water_level_filenames = [line[1], ]
 
                 for index, water_level_filename in enumerate(water_level_filenames):
-                    water_level_tempdir = tempfile.mktemp()
+                    water_level_tempdir = tempfile.mkdtemp()
                     myzip.extract(water_level_filename, water_level_tempdir)
                     tempfilename = os.path.join(water_level_tempdir, water_level_filename)
                     with open(tempfilename) as water_level_tempfile:
@@ -249,6 +258,22 @@ def analyze_zip_file(zipfile):
     return '\n'.join(result)
 
 
+def create_benefit_scenario(all_form_data):
+    """batenkaart"""
+    benefit_scenario = BenefitScenario(
+        name=all_form_data['name'],
+        email=all_form_data['email'],
+        zip_risk_a=all_form_data['zipfile_risk_before'],
+        zip_risk_b=all_form_data['zipfile_risk_after'],)
+    benefit_scenario.save()
+    return benefit_scenario
+
+
+def analyze_benefit_files(zipfile_before, zipfile_after):
+    result = ['start analyse']
+    return '\n'.join(result)
+
+
 class Wizard(ViewContextMixin, SessionWizardView):
     template_name = 'lizard_damage/base_form.html'
     file_storage = temp_storage
@@ -263,7 +288,8 @@ class Wizard(ViewContextMixin, SessionWizardView):
     }
 
     def get_form_initial(self, step):
-        if step == '7':
+        # For batch processing
+        if step == '8':
             form_datas = {}
             for form_step in ('3', '4', '5'):
                 form_data = self.get_cleaned_data_for_step(form_step)
@@ -274,6 +300,15 @@ class Wizard(ViewContextMixin, SessionWizardView):
                 return {'zip_content': analyze_zip_file(form_datas['zipfile'])}
             except:
                 return {'zip_content': 'analyse gefaald, zipfile is niet goed'}
+        # For batenkaart
+        if step == '9':
+            form_data = self.get_cleaned_data_for_step('7')
+            try:
+                return {'zip_content': analyze_benefit_files(
+                    form_data['zipfile_risk_before'],
+                    form_data['zipfile_risk_after'])}
+            except:
+                return {'zip_content': 'analyse gefaald, batenkaart bestanden zijn niet goed'}
         return super(Wizard, self).get_form_initial(step)
 
     # def get_form(self, step=None, data=None, files=None):
@@ -304,13 +339,16 @@ class Wizard(ViewContextMixin, SessionWizardView):
         #do_something_with_the_form_data(form_list)
 
         all_form_data = self.get_all_cleaned_data()
-        print all_form_data
 
         scenario_type = int(all_form_data['scenario_type'])
-        damage_scenario = self.SCENARIO_TYPE_FUNCTIONS[scenario_type](all_form_data)
-
-        # launch task
-        tasks.damage_scenario_to_task(damage_scenario, username="web")
+        if scenario_type in (0, 1, 2, 3, 4, 5):
+            damage_scenario = self.SCENARIO_TYPE_FUNCTIONS[scenario_type](all_form_data)
+            # launch task
+            tasks.damage_scenario_to_task(damage_scenario, username="web")
+        elif scenario_type == 6:
+            # baten taak
+            benefit_scenario = create_benefit_scenario(all_form_data)
+            tasks.benefit_scenario_to_task(benefit_scenario, username="web")
 
         # e-mail received: let's not do this. Feedback is given directly
         # subject = 'Schademodule: Scenario %s ontvangen' % damage_scenario.name
@@ -324,7 +362,7 @@ class DamageScenarioResult(ViewContextMixin, TemplateView):
     template_name = 'lizard_damage/damage_scenario_result.html'
 
     def title(self):
-        return 'Schademodule resultatenpagina %s' % str(self.damage_scenario)
+        return 'STOWA Schade Calculator resultatenpagina %s' % str(self.damage_scenario)
 
     def version(self):
         return tools.version()
@@ -334,7 +372,7 @@ class DamageScenarioResult(ViewContextMixin, TemplateView):
         try:
             root_url = 'http://%s' % Site.objects.all()[0].domain
         except:
-            root_url = 'http://damage.lizard.net'
+            root_url = 'http://schade.lizard.net'
         return root_url
 
     @property
@@ -345,16 +383,120 @@ class DamageScenarioResult(ViewContextMixin, TemplateView):
 class DamageEventKML(ViewContextMixin, TemplateView):
     template_name = 'lizard_damage/event_result.kml'
 
+    # @property
+    # def legend_url(self):
+    #     # Url to an image, optional
     @property
-    def damage_event(self):
-        return get_object_or_404(DamageEvent, slug=self.kwargs['slug'])
+    def legend_url(self):
+        return self.root_url + '/static_media/lizard_damage/legend.png'
+
+    @property
+    def events(self):
+        event = get_object_or_404(DamageEvent, slug=self.kwargs['slug'])
+        return event.damageeventresult_set.all()
 
     @property
     def root_url(self):
         try:
             root_url = 'http://%s' % Site.objects.all()[0].domain
         except:
-            root_url = 'http://damage.lizard.net'
+            root_url = 'http://schade.lizard.net'
+        return root_url
+
+    def get(self, request, *args, **kwargs):
+        context = self.get_context_data(**kwargs)
+        return self.render_to_response(context, mimetype='application/vnd.google-earth.kml+xml')
+
+
+class GeoImageKML(DamageEventKML):
+    @property
+    def events(self):
+        slugs = self.kwargs['slugs']
+        return [get_object_or_404(GeoImage, slug=slug) for slug in slugs.split(',')]
+
+
+class GeoImageLandUseKML(GeoImageKML):
+    @property
+    def legend_url(self):
+        return self.root_url + '/static_media/lizard_damage/legend_landuse.png'
+
+
+class GeoImageHeightKML(GeoImageKML):
+    @property
+    def legend_url(self):
+        """Dirty way to get min/max"""
+        try:
+            #fn something like: height_i43bn2_09_-2230_3249
+            fn = self.kwargs['slugs'].split(',')[0]
+            fn_split = fn.split('_')
+            min_height = float(fn_split[-2]) / 1000
+            max_height = float(fn_split[-1]) / 1000
+        except:
+            min_height = 0.0
+            max_height = 1.0
+        return '%s%s?min_height=%.3f&max_height=%.3f' % (
+            self.root_url, reverse('lizard_damage_legend_height'),
+            min_height, max_height)
+
+
+class LegendHeight(View):
+    """Use pil to take an existing image and add some text on top.
+
+    Note that it uses the STATIC_ROOT folder, so you should run
+    "collectstatic" (which is standard on servers).
+    """
+    def get(self, request, *args, **kwargs):
+        f1 = float(request.GET.get('min_height', '0.0'))
+        f2 = float(request.GET.get('max_height', '1.0'))
+
+        image = Image.open(os.path.join(settings.STATIC_ROOT, "lizard_damage/legend_height.png"))
+        f = ImageFont.load_default()
+
+        draw = ImageDraw.Draw(image)
+        draw.text( (45, 15), " %.1f mNAP" % f2, font=f, fill=(90, 90, 90))
+        draw.text( (45, 65), " %.1f mNAP" % f1, font=f, fill=(90, 90, 90))
+
+        # serialize to HTTP response
+        response = HttpResponse(mimetype="image/png")
+        image.save(response, "PNG")
+        return response
+
+
+class BenefitScenarioResult(ViewContextMixin, TemplateView):
+    template_name = 'lizard_damage/benefit_scenario_result.html'
+
+    def title(self):
+        return 'STOWA Schade Calculator resultatenpagina baten %s' % str(self.benefit_scenario)
+
+    def version(self):
+        return tools.version()
+
+    @property
+    def root_url(self):
+        try:
+            root_url = 'http://%s' % Site.objects.all()[0].domain
+        except:
+            root_url = 'http://schade.lizard.net'
+        return root_url
+
+    @property
+    def benefit_scenario(self):
+        return get_object_or_404(BenefitScenario, slug=self.kwargs['slug'])
+
+
+class BenefitScenarioKML(ViewContextMixin, TemplateView):
+    template_name = 'lizard_damage/benefit_scenario.kml'
+
+    @property
+    def benefit_scenario(self):
+        return get_object_or_404(BenefitScenario, slug=self.kwargs['slug'])
+
+    @property
+    def root_url(self):
+        try:
+            root_url = 'http://%s' % Site.objects.all()[0].domain
+        except:
+            root_url = 'http://schade.lizard.net'
         return root_url
 
     def get(self, request, *args, **kwargs):